+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - Execute unit tests on Travis
 - The highest definition is now selected by default instead of the lowest one for the HTML player
 
+## BUG FIXES
+- Fix the removed first slide when video's beginning is offsetted
+
 ## DEPENDENCIES
 
 - **karma-phantomjs-launcher** has been removed